Fix potential race in generateStackTrace
We ran into an error within Bugsnag:

    panic runtime error: index out of range [8] with length 8
        github.com/bugsnag/bugsnag-go/v2/event.go:219 generateStacktrace
        github.com/bugsnag/bugsnag-go/v2/payload.go:158 (*payload).exceptions
        github.com/bugsnag/bugsnag-go/v2/payload.go:82 (*payload).MarshalJSON
        github.com/bugsnag/bugsnag-go/v2/payload.go:37 (*payload).deliver
        github.com/bugsnag/bugsnag-go/v2/report_publisher.go:23 (*defaultReportPublisher).delivery
        github.com/bugsnag/bugsnag-go/v2/bugsnag.go:53 Configure in goroutine 1

Inspection prompted our engineers to suspect the double call to
`err.StackFrames()` within this function.

This untested commit simply calls the function once, hopefully
bypassing whatever race seems to have tripped us up.

Separately, we suspect that we are configuring Bugsnag strangely within
one of our services; and we intend to fix that as well, as a
"belt-and-suspenders" approach to handling this issue within our
system.

This seems like a safe patch to put in place for ourselves though.

----

We have deliberately not included tests for this. AI tooling did
produce a test, but I don't know whether it's a good test or not.